Update to bullet version from 2020-02-13
https://github.com/bulletphysics/bullet3/tree/b06cdd5fc1328e5dcebedfd351ad72bfe0119cf9

+++ extern/bullet2/src/BulletSoftBody/btSoftRigidDynamicsWorld.cpp
@@ -117,7 +117,7 @@ void btSoftRigidDynamicsWorld::solveSoftBodiesConstraints(btScalar timeStep)
 	m_softBodySolver->solveConstraints(timeStep * m_softBodySolver->getTimeScale());
 }
 
-void btSoftRigidDynamicsWorld::addSoftBody(btSoftBody* body, short int collisionFilterGroup, short int collisionFilterMask)
+void btSoftRigidDynamicsWorld::addSoftBody(btSoftBody* body, int collisionFilterGroup, int collisionFilterMask)
 {
 	m_softBodies.push_back(body);
